+++ packages/lesswrong/lib/components.js
@@ -117,7 +117,7 @@ import '../components/posts/PostsSingle.jsx';
 import '../components/posts/PostsNoMore.jsx';
 import '../components/posts/PostsNoResults.jsx';
 import '../components/posts/PostsLoading.jsx';
-import '../components/posts/PostsDailyList.jsx';
+import '../components/posts/PostsTimeframeList.jsx';
 import '../components/posts/AllPostsPage.jsx';
 import '../components/posts/PostsListSettings.jsx';
 import '../components/posts/PostsThumbnail.jsx';
@@ -142,7 +142,7 @@ import '../components/posts/PostsSingleSlug.jsx';
 import '../components/posts/PostsSingleRoute.jsx';
 import '../components/posts/PostsSingleSlugWrapper.jsx';
 import '../components/posts/PostsList2.jsx';
-import '../components/posts/PostsDay.jsx';
+import '../components/posts/PostsTimeBlock.jsx';
 import '../components/posts/PostsCommentsThread.jsx';
 import '../components/posts/PostsNewForm.jsx';
 import '../components/posts/PostsEditForm.jsx';
